requirements.txt
alembic==1.15.2
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
Brotli==1.1.0
cachelib==0.9.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
cssselect2==0.8.0
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Faker==37.11.0
fastapi==0.123.9
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
filelock==3.19.1
Flask==3.0.3
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.57.0
fsspec==2025.12.0
greenlet==3.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.14
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6
nodeenv==1.9.1
numpy==2.2.6
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.4.0
pre_commit==4.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pypdf==6.0.0
pyphen==0.17.2
pystray==0.19.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
reportlab==4.4.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.47.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.40
starlette==0.50.0
svglib==1.5.1
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.9.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.34.0
watchfiles==1.1.1
weasyprint==62.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.1.2
xhtml2pdf==0.2.17
xlsxwriter==3.2.5
zopfli==0.2.3.post1
